NotifyPeriod changed from long to real: * modified all required objects * still there might be issues due to refactoring

+++ app/src/main/java/pl/projektorion/krzysztof/blesensortag/DBSelectStethoscopeParamRootActivity.java
@@ -38,7 +38,7 @@ public class DBSelectStethoscopeParamRootActivity extends Activity {
                     .getData(DBSelectStethoscopeParamRootData.ATTRIBUTE_ID_RECORD);
             final long dateSeconds = (long) record
                     .getData(DBSelectStethoscopeParamRootData.ATTRIBUTE_DATE_SECONDS);
-            final long notifyPeriod = (long) record
+            final double notifyPeriod = (double) record
                     .getData(DBSelectStethoscopeParamRootData.ATTRIBUTE_NOTIFY_PERIOD);
 
             final DBSelectInterface rootRecord = new DBSelectRootRecordData(_id, dateSeconds);
