+++ app/src/main/java/pl/projektorion/krzysztof/blesensortag/database/inserts/sensors/Barometer/DBInsertBarometerParam.java
@@ -19,7 +19,7 @@ public class DBInsertBarometerParam extends DBInsertParamAbstract {
     @Override
     protected ContentValues values(DBParamDataInterface data) {
         ContentValues values = new ContentValues();
-        final int period = (Integer)
+        final double period = (double)
                 data.getData(DBParamDataInterface.NOTIFY_INTERVAL_PARAMETER);
         values.put(DBTableBarometerParam.NOTIFY_INTERVAL, period);
         values.put(DBTableBarometerParam.COLUMN_ROOT_REF, dbWriter.getRootRowId());
