+++ app/src/main/java/pl/projektorion/krzysztof/blesensortag/database/inserts/sensors/IRTemperature/DBInsertIRTemperatureParam.java
@@ -20,7 +20,7 @@ public class DBInsertIRTemperatureParam extends DBInsertParamAbstract {
     @Override
     protected ContentValues values(DBParamDataInterface data) {
         ContentValues values = new ContentValues();
-        final int period = (Integer)
+        final double period = (double)
                 data.getData(DBParamDataInterface.NOTIFY_INTERVAL_PARAMETER);
         values.put(DBTableIRTemperatureParam.NOTIFY_INTERVAL, period);
         values.put(DBTableIRTemperatureParam.COLUMN_ROOT_REF, dbWriter.getRootRowId());
