+++ app/src/main/java/pl/projektorion/krzysztof/blesensortag/database/selects/sensors/Humidity/DBSelectHumidity.java
@@ -36,7 +36,7 @@ public class DBSelectHumidity extends DBSelectSensorAbstract {
     }
 
     @Override
-    protected void parse_cursor_data(long time, Cursor cursor) {
+    protected void parse_cursor_data(double time, Cursor cursor) {
         humidityRecords.add(new DBSelectHumidityData(time, cursor));
     }
