+++ app/src/main/java/pl/projektorion/krzysztof/blesensortag/database/selects/sensors/Movement/DBSelectMovement.java
@@ -36,7 +36,7 @@ public class DBSelectMovement extends DBSelectSensorAbstract {
     }
 
     @Override
-    protected void parse_cursor_data(long time, Cursor cursor) {
+    protected void parse_cursor_data(double time, Cursor cursor) {
         movementRecords.add(new DBSelectMovementData(time, cursor));
     }
