+++ app/src/main/java/pl/projektorion/krzysztof/blesensortag/database/selects/sensors/OpticalSensor/DBSelectOpticalSensor.java
@@ -36,7 +36,7 @@ public class DBSelectOpticalSensor extends DBSelectSensorAbstract {
     }
 
     @Override
-    protected void parse_cursor_data(long time, Cursor cursor) {
+    protected void parse_cursor_data(double time, Cursor cursor) {
         opticalSensorRecords.add(new DBSelectOpticalSensorData(time, cursor));
     }
